Add ALU output for whether instruction was branch and if output is an addr
Added some logic to ALU to output whether the instruction issued was a
branch instruction. This will be needed by the Execute model to pass on
the "branch taken" bit to stages who need it.

Also added an output Bool to ALU to indicate whether the result is an
address. This will be stored in the RobWriteback structure. It will be
used by the ROB to decide whether the 64-bit value can be ignored, and
used by LSQ to get their computed addresses.