Add support for Icache request cancellation incase of mispredictions
This commit adds support for cancelling previous requests issued to Icache.
This is needed incase of mispredictions where the previous requests must be
invalidated for correctness.

The Fetch module sends a kill signal to the Icache and sends the correct target
in the next cycle. When the Icache receives a kill signal, it invalidates
everything related to s1_* and s2_* and cancels any issued memory request.

Basically there is a one cycle bubble. As soon as we see a misprediction, we
pick the correct target in the next cycle and not in the same cycle