Don't connect SeqMem output to register directly.
This results in a 2 cycle latency resulting in mayhem all over. We get the
output of SeqMem in wires after 1 cycle. Use them to make decisions about hit,
miss etc.  and latch them at the end of 1st cycle.

If we connect the SeqMem output to a register and use the register wires for
decision making, we get output delayed by 1 cycle